+++ src/main/java/exception/EkkoException.java
@@ -1,0 +1,7 @@
+package exception;
+
+public class EkkoException extends Exception {
+    public EkkoException(String message) {
+        super(message);
+    }
+}
