+++ src/main/java/task/Category.java
@@ -1,0 +1,18 @@
+package task;
+
+public enum Category {
+    TODO("T"),
+    DEADLINE("D"),
+    EVENT("E");
+
+    private final String symbol;
+
+    Category(String symbol) {
+        this.symbol = symbol;
+    }
+
+    @Override
+    public String toString() {
+        return symbol;
+    }
+}
