+++ src/main/java/task/Event.java
@@ -1,0 +1,25 @@
+package task;
+
+public class Event extends Task {
+    protected String from;
+    protected String to;
+
+    public Event(String description, String from, String to) {
+        super(description, Category.EVENT);
+        this.from = from;
+        this.to = to;
+    }
+
+    public String getFrom() {
+        return from;
+    }
+
+    public String getTo() {
+        return to;
+    }
+
+    @Override
+    public String toString() {
+        return this.description + " (from: " + from + " to: " + to + ")";
+    }
+}
