+++ src/main/java/task/ToDo.java
@@ -1,0 +1,7 @@
+package task;
+
+public class ToDo extends Task {
+    public ToDo(String description) {
+        super(description, Category.TODO);
+    }
+}
